# [INTEL MKL] Fix dequantize op regression issue