To support the target type of bfloat16, the dequantize_op always transfer the int8 data into the fp32 then cast to target type of bf16 or fp32.
If the target type is FP32, we can skip the cast operation in this PR.