@leslie-fang-intel 2020-05-07
Here is the op's benchmark results:

Before:
```
Benchmark                                    Time(ns) Iterations
----------------------------------------------------------------
BM_DequantizeMinCombinedCpuQuint16             364538       1950         6172.2MB/s 0.0M items/s
BM_DequantizeMinCombinedCpuQint16              374539       1897         6007.4MB/s 0.0M items/s
BM_DequantizeMinCombinedCpuQuint8              351951       1980         5327.5MB/s 0.0M items/s
BM_DequantizeMinCombinedCpuQint8               356293       1958         5262.5MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQuint16     709645        961         2113.7MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQint16      717319        943         2091.1MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQuint8      708349        970         1588.2MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQint8       706389        972         1592.6MB/s 0.0M items/s
```
After fix:
```
Benchmark                                    Time(ns) Iterations
----------------------------------------------------------------
BM_DequantizeMinCombinedCpuQuint16             203394       3470         11062.3MB/s 0.0M items/s
BM_DequantizeMinCombinedCpuQint16              197711       3411         11380.3MB/s 0.0M items/s
BM_DequantizeMinCombinedCpuQuint8              197168       3516         9509.7MB/s 0.0M items/s
BM_DequantizeMinCombinedCpuQint8               197776       3523         9480.4MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQuint16     704444        976         2129.3MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQint16      705720        968         2125.5MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQuint8      699498        981         1608.3MB/s 0.0M items/s
BM_DequantizeBfloat16MinCombinedCpuQint8       702688       1000         1601.0MB/s 0.0M items/s
```